backend/pyproject.toml
[project]
name = "d4-ticket-ai-backend"
version = "0.1.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0,<0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "boto3>=1.35.0",
    "aiobotocore>=2.13.0",
    "pillow>=11.0.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
